Rename renamed import selectors.
Names that are renamed in an import can now be renamed in the file. This is done
using the existing linked-mode-ui rename, with a special case for import
selectors. The code to decide whether a file-local rename should be performed was
duplicated in scala-refactoring, so we now just reuse that.

Fixes #1000884